Update jest and configure CI postgres, get tests passing
