+++ .github/workflows/test.yml
@@ -14,12 +14,14 @@ jobs:
       postgres:
         image: postgres:14
         env:
-          POSTGRES_PASSWORD: postgres
+          POSTGRES_PASSWORD: test
         options: >-
           --health-cmd pg_isready
           --health-interval 10s
           --health-timeout 5s
           --health-retries 5
+        ports:
+          - 5432:5432
 
     steps:
       - uses: actions/checkout@v3
@@ -27,5 +29,9 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
-      - run: npm ci
-      - run: (npm run test -- --coverage --coverageReporters=text-lcov | ./node_modules/.bin/coveralls; exit ${PIPESTATUS[0]})
+      - name: Install dependencies
+        run: npm ci
+      - name: Run tests
+        run: npm run test
+        env:
+          POSTGRES_URL: "postgres://postgres:test@127.0.0.1:5432"
