--- lerna.json
@@ -1,15 +1,0 @@
-{
-  "version": "0.0.45",
-  "packages": [
-    "packages/*"
-  ],
-  "ci": false,
-  "command": {
-    "bootstrap": {
-      "hoist": true,
-      "npmClientArgs": [
-        "--no-package-lock"
-      ]
-    }
-  }
-}
