+++ package.json
@@ -6,17 +6,23 @@
     "node": ">=8.0.0"
   },
   "scripts": {
-    "test": "jest test/*",
+    "test": "jest",
     "test:db": "jest test/database/**"
   },
   "workspaces": [
     "packages/*"
   ],
   "devDependencies": {
+    "@swc/core": "^1.3.23",
+    "@swc/jest": "^0.2.24",
+    "@tsconfig/node-lts": "^18.12.1",
+    "@tsconfig/recommended": "^1.0.1",
+    "@types/jest": "^29.2.4",
     "coveralls": "^3.0.2",
-    "jest": "^23.4.1",
-    "pg": "^7.5.0",
-    "prettier": "^1.14.2"
+    "jest": "^29.3.1",
+    "pg": "^8.8.0",
+    "prettier": "^1.14.2",
+    "typescript": "^4.9.4"
   },
   "volta": {
     "node": "18.12.1"
